Moved the content of the About section to the center of the column for better alignment.
Updated the About section to introduce Luga|Talk, highlighting its purpose and the fusion of Luga-flow with Luganda culture.

Added a horizontal rule with a width of 70% to visually separate the content in the About section.

Applied custom styling to the About section for improved visual appeal, including border, padding, and rounded corners.

Improved the look of the Sign Up and Login buttons by changing their Bootstrap classes.