Added SummernoteWidget to BlogPostForm and create BlogPostForm for the logged in user to create a new post
Added slugify method to the Blogpost model so that the user is not required to enter this field

Made the Feature_image optional to the BlogPost model when the user is creating a blogpost

applied the URL pattern to access the link to create a new blog post.

Added blog post creation view with login requirement and ensuring that on submiting the blog post, the user is redirected to the home page.

Created html code that will be the provide the inteface for the user on creating a new post

Applied the link for the new post to the base template, the link that the user will follow to create a post.

Added a home link to the message when the user is logged