Add user blogposts and like functionality
Created user blogposts file and added code to display a logged-in user's created and liked blogposts. Applied a heart icon in the blogpost HTML file for liking functionality. Created a link on the profile.html file for users to access their blog posts and liked posts. Added views and URL patterns for managing blog posts and likes. Created Like model with relationships to BlogPost and User.